third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/non-anchored-dialog-positioning-expected.txt

Tests positioning of non-anchored dialogs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


showModal() should center in the viewport.
PASS dialog.getBoundingClientRect().top is centeredTop

The computed top and bottom of a centered dialog should still have position auto
PASS window.getComputedStyle(dialog).top is "auto"
PASS window.getComputedStyle(dialog).bottom is "auto"

Dialog should be recentered if showModal() is called after close().
PASS dialog.getBoundingClientRect().top is centeredTop

Dialog should not be recentered on a relayout.
PASS dialog.getBoundingClientRect().top is expectedTop

A tall dialog should be positioned at the top of the viewport.
PASS dialog.getBoundingClientRect().top is 0

The dialog should be centered regardless of the presence of a horizontal scrollbar.
PASS dialog.getBoundingClientRect().top is centeredTop

Test that centering works when dialog is inside positioned containers.
PASS dialog.getBoundingClientRect().top is centeredTop
PASS dialog.getBoundingClientRect().top is centeredTop

Dialog's position should survive after becoming display:none temporarily.
PASS dialog.getBoundingClientRect().top is expectedTop

Dialog loses centering position when removed from document.
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top

Dialog's specified position should survive after close() and showModal().
PASS dialog.getBoundingClientRect().top is expectedTop

Dialog is recentered if showModal() is called after removing 'open'
PASS dialog.getBoundingClientRect().top is centeredTop

Dialog should not be centered if showModal() was called when an ancestor had display 'none'.
PASS dialog.getBoundingClientRect().top is expectedTop

Test setting 'top' on an abspos modal dialog
PASS dialog.getBoundingClientRect().top + window.scrollY is offset

Test setting 'bottom' on an abspos modal dialog
PASS dialog.getBoundingClientRect().bottom + window.scrollY is window.innerHeight - offset

Test fixed position for a modal dialog
PASS dialog.getBoundingClientRect().top is offset

Test absolute position for a non-modal dialog
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top + offset

Test static position for a non-modal dialog
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top

Test relative position for a non-modal dialog
PASS dialog.getBoundingClientRect().top is relativeContainer.getBoundingClientRect().top + offset
PASS dialog.getBoundingClientRect().left is relativeContainer.getBoundingClientRect().left + offset
PASS successfullyParsed is true

TEST COMPLETE
